refactored

+++ DecimalComparator/src/com/marty/Main.java
@@ -7,10 +7,10 @@ public class Main {
     }
 
     public static boolean areEqualByThreeDecimalPlaces(double num1, double num2) {
-        num1 = num1 * 1000;
-        num2 = num2 * 1000;
+        num1 = (int) (num1 * 1000);
+        num2 = (int) (num2 * 1000);
 
-        if ((int) num1 == (int) num2) {
+        if (num1 == num2) {
             return true;
         }
         return false;
